+++ source/Component/Graphics/2d/Sprites/Object/ObjectSprite.c
@@ -96,7 +96,7 @@ void ObjectSprite::destructor()
 
 //——————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————————
 
-bool ObjectSprite::onEvent(ListenerObject eventFirer __attribute__((unused)), uint32 eventCode)
+bool ObjectSprite::onEvent(ListenerObject eventFirer __attribute__((unused)), uint16 eventCode)
 {
 	switch(eventCode)
 	{
